+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopkeeper/player/book/BookPlayerShopType.java
@@ -4,11 +4,7 @@ import java.util.Collections;
 import java.util.List;
 
 import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
-import com.nisovin.shopkeepers.api.shopkeeper.ShopCreationData;
-import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperCreateException;
-import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopCreationData;
 import com.nisovin.shopkeepers.lang.Messages;
-import com.nisovin.shopkeepers.shopkeeper.ShopkeeperData;
 import com.nisovin.shopkeepers.shopkeeper.player.AbstractPlayerShopType;
 
 public final class BookPlayerShopType extends AbstractPlayerShopType<SKBookPlayerShopkeeper> {
@@ -39,15 +35,7 @@ public final class BookPlayerShopType extends AbstractPlayerShopType<SKBookPlaye
 	}
 
 	@Override
-	public SKBookPlayerShopkeeper createShopkeeper(int id, ShopCreationData shopCreationData) throws ShopkeeperCreateException {
-		this.validateCreationData(shopCreationData);
-		SKBookPlayerShopkeeper shopkeeper = new SKBookPlayerShopkeeper(id, (PlayerShopCreationData) shopCreationData);
-		return shopkeeper;
-	}
-
-	@Override
-	public SKBookPlayerShopkeeper loadShopkeeper(ShopkeeperData shopkeeperData) throws ShopkeeperCreateException {
-		this.validateShopkeeperData(shopkeeperData);
-		return new SKBookPlayerShopkeeper(shopkeeperData);
+	protected SKBookPlayerShopkeeper createNewShopkeeper() {
+		return new SKBookPlayerShopkeeper();
 	}
 }