+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopkeeper/player/trade/TradingPlayerShopType.java
@@ -4,11 +4,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
-import com.nisovin.shopkeepers.api.shopkeeper.ShopCreationData;
-import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperCreateException;
-import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopCreationData;
 import com.nisovin.shopkeepers.lang.Messages;
-import com.nisovin.shopkeepers.shopkeeper.ShopkeeperData;
 import com.nisovin.shopkeepers.shopkeeper.player.AbstractPlayerShopType;
 
 public final class TradingPlayerShopType extends AbstractPlayerShopType<SKTradingPlayerShopkeeper> {
@@ -39,15 +35,7 @@ public final class TradingPlayerShopType extends AbstractPlayerShopType<SKTradin
 	}
 
 	@Override
-	public SKTradingPlayerShopkeeper createShopkeeper(int id, ShopCreationData shopCreationData) throws ShopkeeperCreateException {
-		this.validateCreationData(shopCreationData);
-		SKTradingPlayerShopkeeper shopkeeper = new SKTradingPlayerShopkeeper(id, (PlayerShopCreationData) shopCreationData);
-		return shopkeeper;
-	}
-
-	@Override
-	public SKTradingPlayerShopkeeper loadShopkeeper(ShopkeeperData shopkeeperData) throws ShopkeeperCreateException {
-		this.validateShopkeeperData(shopkeeperData);
-		return new SKTradingPlayerShopkeeper(shopkeeperData);
+	protected SKTradingPlayerShopkeeper createNewShopkeeper() {
+		return new SKTradingPlayerShopkeeper();
 	}
 }